requirements.txt
pyserial
pyyaml
streamlit
plotly
pandas
